Use background loading for all requests
