+++ lib/background.js
@@ -1,44 +1,37 @@
-function makeRequest(method, url) {
-  return new Promise(function (resolve, reject) {
-    const xhr = new XMLHttpRequest();
-    xhr.open(method, url);
+import $ from 'jquery';
 
-    xhr.onload = function () {
-      if (this.status >= 200 && this.status < 300) {
-        if (xhr.response) {
-          return resolve(JSON.parse(xhr.response));
-        }
+function loader(urls, cb) {
+  const req = urls.shift();
+  const url = req.url || req;
 
-        resolve();
-      } else {
-        reject({
-          status: this.status,
-          statusText: xhr.statusText,
-        });
-      }
-    };
-    xhr.onerror = function () {
-      reject({
-        status: this.status,
-        statusText: xhr.statusText,
-      });
-    };
+  $.ajax({
+    method: req.method || 'HEAD',
+    url,
+  }).then(function (res) {
+    cb(null, url, res);
+  }).fail(function () {
+    if (urls.length === 0) {
+      return cb(new Error('Could not load any url'));
+    }
 
-    xhr.send();
+    loader(urls, cb);
   });
 }
 
-chrome.runtime.onMessage.addListener(({ method, url }, sender) => {
-  makeRequest(method, url)
-    .then((body) => {
-      chrome.tabs.sendMessage(sender.tab.id, {
-        err: null,
-        body,
-      });
-    })
-    .catch((err) => {
+function init() {
+  if (chrome.runtime.onMessage.hasListeners()) {
+    return;
+  }
+
+  chrome.runtime.onMessage.addListener(({ urls }, sender) => {
+    loader(urls, (err, url, res) => {
       chrome.tabs.sendMessage(sender.tab.id, {
         err,
+        url,
+        res,
       });
     });
-});
+  });
+}
+
+init();
