+++ lib/click-handler.js
@@ -1,5 +1,4 @@
 import $ from 'jquery';
-import tryLoad from '../packages/try-load';
 import { showTooltip } from './gh-interface.js';
 import * as resolvers from './resolver/index.js';
 
@@ -17,6 +16,24 @@ function openUrl(url, newWindow = false) {
   }
 
   window.open(url, newWindow ? '_blank' : '_self');
+}
+
+function tryLoad(urls, cb) {
+  chrome.runtime.sendMessage({
+    urls,
+  });
+
+  if (chrome.runtime.onMessage.hasListeners()) {
+    return;
+  }
+
+  chrome.runtime.onMessage.addListener((msg) => {
+    if (msg.err) {
+      return cb(new Error('Could not load any url'));
+    }
+
+    cb(null, msg.url, msg.res);
+  });
 }
 
 function getResolverUrls(dataAttr) {
